requirements.txt
Django>=3.2.4,<3.3
djangorestframework>=3.12.4,<3.13
mysqlclient>=2.0.3,<2.1
drf-spectacular>=0.15.1,<0.16
djangorestframework-simplejwt==4.8.0
PyJWT==2.3.0
django-redis>=4.12.1,<4.13
redis>=3.5.3,<4.0